Implement quest starting, completing, and forfeiting. This includes validation and rewarding of all pre-quest and post-quest requirements/rewards. Attempting to complete a quest should give the most relevant error possible. (E.g. If your inventory is too full, it should say so.) There are still many unimplemented pieces of quests, so don't expect them to work fully yet. Item code is also optimized in certain spots.

+++ src/Maple/Quest.cs
@@ -125,10 +125,10 @@ namespace Destiny.Maple
 
             this.PreRequiredQuests  = requests.Where(x => (string)x["request_type"] == "quest" && (string)x["quest_state"] == "start")
                                               .Select(x => new { objectid = (int)x["objectid"], quantity = (short)x["quantity"] })
-                                              .ToDictionary(x => (ushort)x.objectid, x => (QuestStatus)x.quantity);
+                                              .ToDictionary(x => (ushort)x.objectid, x => (QuestStatus)(byte)x.quantity);
             this.PostRequiredQuests = requests.Where(x => (string)x["request_type"] == "quest" && (string)x["quest_state"] == "end")
                                               .Select(x => new { objectid = (int)x["objectid"], quantity = (short)x["quantity"] })
-                                              .ToDictionary(x => (ushort)x.objectid, x => (QuestStatus)x.quantity);
+                                              .ToDictionary(x => (ushort)x.objectid, x => (QuestStatus)(byte)x.quantity);
             this.PreRequiredItems   = requests.Where(x => (string)x["request_type"] == "item" && (string)x["quest_state"] == "start")
                                               .Select(x => new { objectid = (int)x["objectid"], quantity = (short)x["quantity"] })
                                               .ToDictionary(x => x.objectid, x => x.quantity);
